fixed armduino and other stuff

+++ ArmStuff/Armduino/Motor.h
@@ -1,6 +1,7 @@
 #ifndef Motor_h
 #define Motor_h
 #include "Arduino.h"
+#include <Servo.h>
 #include "Servo.h"
 #include "ros.h";
 #include <std_msgs/UInt16.h>
@@ -10,7 +11,7 @@ class Motor{
     Motor(int _pin, int _dirPin);
     Motor(int _pin);
     Motor();
-    void doMotor(uint16_t _drive);
+    void doMotor(int _drive);
     void doMotorsPWM(std_msgs::Float32 _input);
   private:
     int pin;
